Python_Medical_Insurance_Project
In this project we are analyzing a dataset from Kaggle (https://www.kaggle.com/datasets/mirichoi0218/insurance) we are aimming for a causal analysis, by looking first for the region with the greatest insurance charges and then analyzing per region the risk factors that are making this cost increase, we make use of dictionaries, list, loops, variables, dataframes from pandas. All the information comes from a CSV file.